+++ src/PJ/picdrive/targa/targa.c
@@ -215,7 +215,7 @@ static Errcode read_first(Image_file *ifile, Rcel *screen)
 
 static Errcode save_frames(Image_file	*ifile,
 						   Rcel 		*screen,
-						   int			num_frames,
+						   ULONG		num_frames,
 						   Errcode		(*seek_frame)(int ix,void *seek_data),
 						   void 		*seek_data,
 						   Rcel 		*work_screen )
